refs #552 - Move OSGi Service API from JPA to JakartaPersistence
- renamed project from jpa to jakartapersistence
- moved package from javax to jakarta
- modified spec version
- modified spec names
- modified property names corresponding to new java spec
