+++ cnf/ext/central.mvn
@@ -253,6 +253,7 @@ jakarta.annotation:jakarta.annotation-api:2.1.1
 jakarta.inject:jakarta.inject-api:2.0.1
 jakarta.validation:jakarta.validation-api:3.0.2
 jakarta.xml.bind:jakarta.xml.bind-api:4.0.0
+jakarta.persistence:jakarta.persistence-api:3.0.0
 org.apache.felix:org.apache.felix.http.servlet-api:2.0.0
 
 org.ow2.asm:asm:jar:9.3
