+++ org.osgi.service.jakartapersistence/.classpath
