+++ org.osgi.service.jakartapersistence/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.osgi.service.jpa</name>
+	<name>org.osgi.service.jakartapersistence</name>
 	<comment></comment>
 	<projects>
 	</projects>
