+++ org.osgi.service.jakartapersistence/bnd.bnd
@@ -8,6 +8,6 @@ Export-Package: ${p}.*; -split-package:=first
     org.osgi.framework;maven-scope=provided;version=1.8, \
     org.osgi.resource;maven-scope=provided;version=1.0, \
     org.osgi.namespace.extender;version=1.0.1, \
-    org.eclipse.persistence.jpa_spec;maven-scope=provided;version=2.0;packages=javax.persistence.*
+    jakarta.persistence-api;version='3.0.0'
 
--fixupmessages.jpa_spec: Version for package org\.osgi\.service\.jpa is set to different values
+-contract: JakartaPersistence
